backend_old/requirements.txt
Flask==2.1.3
Flask-RESTful==0.3.9
gunicorn==20.1.0
SQLAlchemy==1.4.25
psycopg2-binary==2.9.1
boto3==1.24.20
python-dotenv==0.20.0
PyPDF2==1.26.0
